Combine GetDevicesTask and GetDevicesResultTask
I initially used separate classes for the "worker thread" runnable and
the "do something with the results on the main thread" runnable because
that is the pattern demonstrated in the MDN documentation [0].

Further code review led me to another common pattern in the codebase,
which is to use a single runnable class for both tasks and choose the
behavior of the runnable based by inspecting the current thread. This is
an improvement becuase it avoids unnecessary boilerplate, code
duplication, and parameter passing.

[0]: https://developer.mozilla.org/en-US/docs/Mozilla/Tech/XPCOM/Guide/Cross-thread_calls_using_runnables