WIP: NativeGetDevices for Mac
Implements the basic IOHIDManager setup and cleanup for the native
MacHidService.

Enumerates HID devices, but does not yet fully construct
nsIHidDeviceInfo for each one because extracting the necessary info from
the Mac native API is somewhat cumbersome, and I am focusing on testing
the multithreaded interactions before writing all of that out.

Problem: I thought I could pass an nsISimpleEnumerator in
NS_NewRunnableFunction to be returned as a parameter to the callback on
the main thread, but I cannot do so because nsISimpleEnumerator is not
thread-safe. At the moment, this code compiles but when
test_hidservice.js is run, it crashes with:

"Hit MOZ_CRASH(nsCOMArrayEnumerator not thread-safe)"