Don't separate Report IDs from Report data in API callbacks
The initial high-level HID API (as defined in the .idls) was modeled on
Chrome's chrome.hid API, which separates Report IDs from Report data in
the send and receive functions. This is slightly more ergonomic for
developers, but requires the HID API to parse report descriptors in
order to determine whether or not a given interface supports multiple
reports or not (see Chromium's `device/hid/hid_report_descriptor.cc`).

I am dropping this separation for now because it requires significant
additional development work for a minor improvement in the developer
ergonomics of the high-level API. In the interest of finishing the HID
API as quickly as possible, I think it makes sense to drop the
separation and require HID API users to be responsible for parsing
report IDs.

We may revisit this decision in the future if it becomes apparent that
the additional development effort to support the separation between
report IDs and report data in the high-level API is worthwhile.