Remove unnecessary XPIDL goo from nsIHid{Connection,DeviceInfo}
I ran into build errors as soon as I added the NativeDisconnect virtual
function to HidConnection. This was because the big blob of cargo-culted
XPIDL goo needed a constructor for NS_GENERIC_FACTORY_CONSTRUCTOR, but
adding virtual methods to HidConnection meant that it could not be
constructed directly.

After some discussion on #developers, I learned that it isn't useful to
construct HidConnection (it's an abstract base classes, only the
platform-specific subclasses should be directly instantiated), and so
the cargo-culted code was unnecessary for HidConnection and
HidDeviceInfo. After removing it, HidConnection has a virtual
NativeDisconnect method, Firefox builds, and my test_hidservice.js test
still passes.

Since I had also identified this as an issue for HidDeviceInfo when I
originally wrote it, but didn't know how to resolve it then, this commit
also removes the unnecessary XPIDL boilerplate from HidDeviceInfo.